transformation de tous les double en float

+++ core/src/fr/limsi/rorqual/core/dpe/enums/chauffageproperties/Chauffage.java
@@ -85,17 +85,17 @@ public class Chauffage {
     private Generateur generateur;
     private Type type;
     private Emission emission;
-    private double prs1;
-    private double rg;
-    private double re;
-    private double rd;
-    private double rr;
-    private double ich;
-    private double pn;
-    private double rpn;
-    private double rpint;
-    private double qp0;
-    private double puissanceVeilleuse;
+    private float prs1;
+    private float rg;
+    private float re;
+    private float rd;
+    private float rr;
+    private float ich;
+    private float pn;
+    private float rpn;
+    private float rpint;
+    private float qp0;
+    private float puissanceVeilleuse;
     private boolean generateurDansVolumeChauffe;
     private boolean presenceRobinetThermostatique;
     private boolean genereEgalementEcs;
@@ -137,7 +137,7 @@ public class Chauffage {
                 case CHAUDIERE_GAZ_CONDENSATION_AVANT_1986:
                 case CHAUDIERE_GAZ_CONDENSATION_ENTRE_1986_ET_2001:
                 case CHAUDIERE_GAZ_CONDENSATION_APRES_2001:
-                    this.prs1 = 3.6;
+                    this.prs1 = 3.6f;
                     break;
                 default:
                     this.prs1 = 0;
@@ -154,14 +154,14 @@ public class Chauffage {
             case POIL_OU_INSERT_BOIS_AVANT_2001:
             case RADIATEUR_GAZ_AVANT_2006:
             case RADIATEUR_GAZ_APRES_2006:
-                this.re=0.95;
+                this.re=0.95f;
                 break;
             case CHAUFFAGE_ELECTRIQUE_DIRECT_RECENT:
-                this.re=0.96;
+                this.re=0.96f;
                 break;
             default:
                 if(this.emission.equals(Emission.RADIATEUR)){
-                    this.re=0.95;
+                    this.re=0.95f;
                     break;
                 }else{
                     this.re=1;
@@ -182,7 +182,7 @@ public class Chauffage {
                 this.rd=1;
                 break;
             default:
-                this.rd=0.9;
+                this.rd=0.9f;
         }
         actualiseIch();
     }
@@ -191,24 +191,24 @@ public class Chauffage {
             case CHAUFFAGE_ELECTRIQUE_DIRECT_ANCIEN:
             case RADIATEUR_GAZ_AVANT_2006:
             case RADIATEUR_GAZ_APRES_2006:
-                this.rr=0.96;
+                this.rr=0.96f;
                 break;
             case CHAUFFAGE_ELECTRIQUE_DIRECT_RECENT:
-                this.rr=0.99;
+                this.rr=0.99f;
                 break;
             case POIL_FIOUL_OU_GPL:
             case POIL_OU_INSERT_BOIS_APRES_2001:
             case POIL_OU_INSERT_BOIS_AVANT_2001:
-                this.rr=0.8;
+                this.rr=0.8f;
                 break;
             default:
                 if(this.emission!=Emission.RADIATEUR){
-                    this.rr=0.96;
+                    this.rr=0.96f;
                 }else{
                     if(this.presenceRobinetThermostatique){
-                        this.rr=0.95;
+                        this.rr=0.95f;
                     }else{
-                        this.rr=0.9;
+                        this.rr=0.9f;
                     }
                 }
         }
@@ -220,23 +220,23 @@ public class Chauffage {
             case CHAUFFAGE_ELECTRIQUE_DIRECT_RECENT:
                 this.rg=1;
             case CHAUDIERE_ELECTRIQUE:
-                this.rg=0.77;
+                this.rg=0.77f;
             case POMPE_A_CHALEUR_AIR_AIR:
-                this.rg=2.2;
+                this.rg=2.2f;
             case POMPE_A_CHALEUR_AIR_EAU:
                 this.rg=2;
             case POMPE_A_CHALEUR_SUR_NAPPE:
-                this.rg=3.2;
+                this.rg=3.2f;
             case POMPE_A_CHALEUR_GEOTHERMIQUE:
                 this.rg=4;
             case POIL_OU_INSERT_BOIS_AVANT_2001:
-                this.rg=0.66;
+                this.rg=0.66f;
             case POIL_OU_INSERT_BOIS_APRES_2001:
-                this.rg=0.78;
+                this.rg=0.78f;
             case POIL_FIOUL_OU_GPL:
-                this.rg=0.72;
+                this.rg=0.72f;
             default:
-                this.rg=0.66;
+                this.rg=0.66f;
         }
         actualiseIch();
     }
@@ -247,16 +247,16 @@ public class Chauffage {
     // Getter
     public Generateur getGenerateur(){return this.generateur;}
     public Type getType(){return this.type;}
-    public double getPrs1() {return this.prs1;}
-    public double getRe() {return this.re;}
-    public double getRd() {return this.rd;}
-    public double getRr() {return this.rr;}
-    public double getQp0() {return qp0;}
-    public double getRpint() {return rpint;}
-    public double getRpn(){ return rpn; }
-    public double getPn() {return pn;}
-    public double getIch() {return ich;}
-    public double getPuissanceVeilleuse() {return puissanceVeilleuse;}
+    public float getPrs1() {return this.prs1;}
+    public float getRe() {return this.re;}
+    public float getRd() {return this.rd;}
+    public float getRr() {return this.rr;}
+    public float getQp0() {return qp0;}
+    public float getRpint() {return rpint;}
+    public float getRpn(){ return rpn; }
+    public float getPn() {return pn;}
+    public float getIch() {return ich;}
+    public float getPuissanceVeilleuse() {return puissanceVeilleuse;}
     public boolean getGenereEgalementEcs(){ return this.genereEgalementEcs; }
 
     // Setter
@@ -273,132 +273,132 @@ public class Chauffage {
         this.presenceRobinetThermostatique=presence;
         this.actualiseRr();
     }
-    public void setPn(double pn) {
+    public void setPn(float pn) {
         this.pn = pn;
         switch(this.generateur){
             case CHAUDIERE_GAZ_CLASSIQUE_AVANT_1981:
-                this.rpn=79+2*Math.log(pn);
-                this.rpint=73+3*Math.log(pn);
-                this.qp0=0.04*pn;
-                this.puissanceVeilleuse=0.24;
+                this.rpn=79+2*(float)Math.log(pn);
+                this.rpint=73+3*(float)Math.log(pn);
+                this.qp0=0.04f*pn;
+                this.puissanceVeilleuse=0.24f;
                 break;
             case CHAUDIERE_GAZ_CLASSIQUE_ENTRE_1981_ET_1986:
-                this.rpn=82+2*Math.log(pn);
-                this.rpint=76+3*Math.log(pn);
-                this.qp0=0.02*pn;
-                this.puissanceVeilleuse=0.15;
+                this.rpn=82+2*(float)Math.log(pn);
+                this.rpint=76+3*(float)Math.log(pn);
+                this.qp0=0.02f*pn;
+                this.puissanceVeilleuse=0.15f;
                 break;
             case CHAUDIERE_GAZ_CLASSIQUE_ENTRE_1986_ET_1991:
-                this.rpn=83+2*Math.log(pn);
-                this.rpint=79+3*Math.log(pn);
-                this.qp0=0.015*pn;
-                this.puissanceVeilleuse=0.15;
+                this.rpn=83+2*(float)Math.log(pn);
+                this.rpint=79+3*(float)Math.log(pn);
+                this.qp0=0.015f*pn;
+                this.puissanceVeilleuse=0.15f;
                 break;
             case CHAUDIERE_GAZ_STANDARD_ENTRE_1991_ET_2001:
-                this.rpn=84+2*Math.log(pn);
-                this.rpint=80+3*Math.log(pn);
-                this.qp0=0.012*pn;
-                this.puissanceVeilleuse=0.12;
+                this.rpn=84+2*(float)Math.log(pn);
+                this.rpint=80+3*(float)Math.log(pn);
+                this.qp0=0.012f*pn;
+                this.puissanceVeilleuse=0.12f;
                 break;
             case CHAUDIERE_GAZ_STANDARD_APRES_2001:
-                this.rpn=84+2*Math.log(pn);
-                this.rpint=80+3*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=84+2*(float)Math.log(pn);
+                this.rpint=80+3*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_GAZ_BASSE_TEMPERATURE_AVANT_2001:
-                this.rpn=87.5+1.5*Math.log(pn);
-                this.rpint=87.5+1.5*Math.log(pn);
-                this.qp0=0.012*pn;
-                this.puissanceVeilleuse=0.12;
+                this.rpn=87.5f+1.5f*(float)Math.log(pn);
+                this.rpint=87.5f+1.5f*(float)Math.log(pn);
+                this.qp0=0.012f*pn;
+                this.puissanceVeilleuse=0.12f;
                 break;
             case CHAUDIERE_GAZ_BASSE_TEMPERATURE_APRES_2001:
-                this.rpn=87.5+1.5*Math.log(pn);
-                this.rpint=87.5+1.5*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=87.5f+1.5f*(float)Math.log(pn);
+                this.rpint=87.5f+1.5f*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_GAZ_CONDENSATION_AVANT_1986:
-                this.rpn=91+1*Math.log(pn);
-                this.rpint=97+1*Math.log(pn);
-                this.qp0=0.01*pn;
-                this.puissanceVeilleuse=0.15;
+                this.rpn=91+1*(float)Math.log(pn);
+                this.rpint=97+1*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
+                this.puissanceVeilleuse=0.15f;
                 break;
             case CHAUDIERE_GAZ_CONDENSATION_ENTRE_1986_ET_2001:
-                this.rpn=91+1*Math.log(pn);
-                this.rpint=97+1*Math.log(pn);
-                this.qp0=0.01*pn;
-                this.puissanceVeilleuse=0.12;
+                this.rpn=91+1*(float)Math.log(pn);
+                this.rpint=97+1*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
+                this.puissanceVeilleuse=0.12f;
                 break;
             case CHAUDIERE_GAZ_CONDENSATION_APRES_2001:
-                this.rpn=91+1*Math.log(pn);
-                this.rpint=97+1*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=91+1*(float)Math.log(pn);
+                this.rpint=97+1*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_CLASSIQUE_AVANT_1970:
-                this.rpn=74+2*Math.log(pn);
-                this.rpint=63+3*Math.log(pn);
-                this.qp0=0.04*pn;
+                this.rpn=74+2*(float)Math.log(pn);
+                this.rpint=63+3*(float)Math.log(pn);
+                this.qp0=0.04f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_CLASSIQUE_ENTRE_1970_ET_1976:
-                this.rpn=77+2*Math.log(pn);
-                this.rpint=71+3*Math.log(pn);
-                this.qp0=0.03*pn;
+                this.rpn=77+2*(float)Math.log(pn);
+                this.rpint=71+3*(float)Math.log(pn);
+                this.qp0=0.03f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_CLASSIQUE_ENTRE_1976_ET_1981:
-                this.rpn=78+2*Math.log(pn);
-                this.rpint=76+3*Math.log(pn);
-                this.qp0=0.02*pn;
+                this.rpn=78+2*(float)Math.log(pn);
+                this.rpint=76+3*(float)Math.log(pn);
+                this.qp0=0.02f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_CLASSIQUE_ENTRE_1981_ET_1991:
-                this.rpn=80+2*Math.log(pn);
-                this.rpint=78+3*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=80+2*(float)Math.log(pn);
+                this.rpint=78+3*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_STANDARD_APRES_1991:
-                this.rpn=84+2*Math.log(pn);
-                this.rpint=80+3*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=84+2*(float)Math.log(pn);
+                this.rpint=80+3*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_BASSE_TEMPERATURE:
-                this.rpn=87.5+1.5*Math.log(pn);
-                this.rpint=87.5+1.5*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=87.5f+1.5f*(float)Math.log(pn);
+                this.rpint=87.5f+1.5f*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_FIOUL_CONDENSATION:
-                this.rpn=91+1*Math.log(pn);
-                this.rpint=97+1*Math.log(pn);
-                this.qp0=0.01*pn;
+                this.rpn=91+1*(float)Math.log(pn);
+                this.rpint=97+1*(float)Math.log(pn);
+                this.qp0=0.01f*pn;
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_BOIS_PLUS_DE_15_ANS:
-                this.rpn=57+6*Math.log(pn);
-                this.rpint=58+6*Math.log(pn);
-                this.qp0=0.07*Math.pow(pn,0.7);
+                this.rpn=57+6*(float)Math.log(pn);
+                this.rpint=58+6*(float)Math.log(pn);
+                this.qp0=0.07f*(float)Math.pow(pn,0.7);
                 this.puissanceVeilleuse=0;
                 break;
             case CHAUDIERE_BOIS_MOINS_DE_15_ANS:
-                this.rpn=80+2*Math.log(pn);
-                this.rpint=81+2*Math.log(pn);
-                this.qp0=0.085*Math.pow(pn,0.6);
+                this.rpn=80+2*(float)Math.log(pn);
+                this.rpint=81+2*(float)Math.log(pn);
+                this.qp0=0.085f*(float)Math.pow(pn,0.6);
                 this.puissanceVeilleuse=0;
                 break;
             case RADIATEUR_GAZ_AVANT_2006:
                 this.rpn=70;
                 break;
             case RADIATEUR_GAZ_APRES_2006:
-                this.rpn=80*Math.log(pn);
+                this.rpn=80*(float)Math.log(pn);
                 break;
         }
     }
-    public void setRg(double rg){
+    public void setRg(float rg){
         this.rg=rg;
         this.actualiseIch();
     }
