+++ core/src/fr/limsi/rorqual/core/dpe/enums/chauffageproperties/FrequenceUtilisationPoilEnum.java
@@ -4,14 +4,14 @@ package fr.limsi.rorqual.core.dpe.enums.chauffageproperties;
  * Created by ricordeau on 21/07/15.
  */
 public enum FrequenceUtilisationPoilEnum {
-    TRES_PEU(0.1),
-    DE_TEMPS_EN_TEMPS(0.25),
-    PRINCIPALEMENT(0.75),
-    QUASI_EXCLUSIVEMENT(0.9);
+    TRES_PEU(0.1f),
+    DE_TEMPS_EN_TEMPS(0.25f),
+    PRINCIPALEMENT(0.75f),
+    QUASI_EXCLUSIVEMENT(0.9f);
 
-    private double k;
-    FrequenceUtilisationPoilEnum(double k){
+    private float k;
+    FrequenceUtilisationPoilEnum(float k){
         this.k=k;
     }
-    public double getFrequence(){ return this.k; }
+    public float getFrequence(){ return this.k; }
 }
